facial-recognition-service/requirements.txt
Flask
python-dotenv
opencv-python
face_recognition
numpy
requests
Pillow